requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.4.4
bootstrap4==0.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.4
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.39.1
django-ckeditor==5.9.0
django-constance==2.7.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-picklefield==3.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.8
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
pycodestyle==2.5.0
pycryptodome==3.9.9
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python3-openid==3.1.0
pytz==2020.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
urllib3==1.24.2
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
